Separate the adapted engine, estinant builder, and odeshin boundaries
