+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/engine-program/engineFunctionConfiguration.ts
@@ -28,7 +28,7 @@ export type CoreEngineFunction2Configuration = {
 
 export type AdaptedEngineFunctionConfiguration = {
   typeName: EngineFunctionConfigurationTypeName.Adapted;
-  filePath: 'packages/voictents-and-estinants-engine/src/adapter/digikikify.ts';
+  filePath: 'packages/voictents-and-estinants-engine/src/adapter/engine/digikikify.ts';
   exportedIdentifier: 'digikikify';
   estinantListKeyIdentifierName: 'estinantTuple';
   explicitVoictentTupleKeyIdentifierName: 'explicitVoictentTuple';
@@ -79,7 +79,7 @@ export const ADAPTED_ENGINE_FUNCTION_CONFIGURATION: AdaptedEngineFunctionConfigu
   {
     typeName: EngineFunctionConfigurationTypeName.Adapted,
     filePath:
-      'packages/voictents-and-estinants-engine/src/adapter/digikikify.ts',
+      'packages/voictents-and-estinants-engine/src/adapter/engine/digikikify.ts',
     exportedIdentifier: 'digikikify',
     estinantListKeyIdentifierName: 'estinantTuple',
     explicitVoictentTupleKeyIdentifierName: 'explicitVoictentTuple',
