+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file-relationships/graph-element/getExternalModuleMetadata.ts
@@ -1,5 +1,5 @@
 import { buildEstinant } from '../../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../../adapter/odeshin/odeshin2';
 import { NodeShape } from '../../graph-visualization/directed-graph/directedGraphNode';
 import { EXTERNAL_BOUNDARY_ZORN } from './boundaryConfiguration';
 import {
