+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file-relationships/graph-element/getFileNodeMetadata.ts
@@ -1,6 +1,6 @@
 import Case from 'case';
 import { buildEstinant } from '../../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../../adapter/odeshin/odeshin2';
 import {
   PROGRAM_ERROR_GEPP,
   ProgramErrorElementLocatorTypeName,
