+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file-relationships/graph-element/getInitialEdgeMetadata.ts
@@ -28,7 +28,7 @@ import {
   ReportedProgramError,
   ReportingEstinantLocator,
 } from '../../error/programError';
-import { OdeshinZorn } from '../../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../../adapter/odeshin/odeshin2';
 
 const ESTINANT_NAME = 'getInitialEdgeMetadata' as const;
 type EstinantName = typeof ESTINANT_NAME;
