+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file/canonical-declaration/canonicalDeclarationOmissionList.ts
@@ -18,9 +18,9 @@ export const canonicalDeclarationOmissionList: LintAssertionOmission[] = [
   'packages/voictents-and-estinants-engine/src/core/engine-shell/estinant/tropoignant.ts',
   'packages/voictents-and-estinants-engine/src/core/internal/platomity.ts',
   'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/estinantBuilder.ts',
-  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/estinantBuilderContext.ts',
-  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/partialAppreffinge.ts',
-  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/vicken.ts',
+  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/shared/estinantBuilderContext.ts',
+  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/shared/partialAppreffinge.ts',
+  'packages/voictents-and-estinants-engine/src/adapter/estinant-builder/shared/vicken.ts',
   'packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/engine-program/input-output/engineEstinantInput2.ts',
   'packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/engine-program/input-output/engineEstinantOutput2.ts',
   'packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/error/programError.ts',
