+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/comments-example/commentsExample.ts
@@ -1,5 +1,8 @@
 import { InMemoryVoictent } from '../../../core/engine/inMemoryVoictent';
-import { buildVoictentByGepp, digikikify } from '../../../adapter/digikikify';
+import {
+  buildVoictentByGepp,
+  digikikify,
+} from '../../../adapter/engine/digikikify';
 import { ProgramFileCache } from '../../../utilities/programFileCache';
 import { ProgramErrorVoictent } from '../../programmable-units/error/programErrorVoictent';
 import { categorizeFiles } from '../../programmable-units/file/categorizeFiles';
