+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/find-unused-exports/findUnusedExports.ts
@@ -1,5 +1,8 @@
 import { InMemoryVoictent } from '../../../core/engine/inMemoryVoictent';
-import { buildVoictentByGepp, digikikify } from '../../../adapter/digikikify';
+import {
+  buildVoictentByGepp,
+  digikikify,
+} from '../../../adapter/engine/digikikify';
 import { ProgramFileCache } from '../../../utilities/programFileCache';
 import { PROGRAM_ERROR_GEPP } from '../../programmable-units/error/programError';
 import { ProgramErrorVoictent } from '../../programmable-units/error/programErrorVoictent';
