+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/model-programs/getEngineEstinantGraphElements.ts
@@ -2,7 +2,7 @@ import { getTextDigest } from '../../../utilities/getTextDigest';
 import { getZorn } from '../../../utilities/getZorn';
 import { isNotNull } from '../../../utilities/isNotNull';
 import { buildEstinant } from '../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../adapter/odeshin/odeshin2';
 import {
   ENGINE_ESTINANT_3_GEPP,
   EngineEstinant3Voque,
