+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/model-programs/getInputEdges.ts
@@ -1,5 +1,5 @@
 import { buildEstinant } from '../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../adapter/odeshin/odeshin2';
 import {
   EngineEstinant3Voque,
   ENGINE_ESTINANT_3_GEPP,
