+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/model-programs/modelPrograms.ts
@@ -1,4 +1,7 @@
-import { buildVoictentByGepp, digikikify } from '../../../adapter/digikikify';
+import {
+  buildVoictentByGepp,
+  digikikify,
+} from '../../../adapter/engine/digikikify';
 import {
   FILE_SYSTEM_OBJECT_ENUMERATOR_CONFIGURATION_GEPP,
   FileSystemObjectEnumeratorConfigurationVoque,
