+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/dependency/getFileDependencies.ts
@@ -1,6 +1,6 @@
 import { NonEmptyTuple } from '../../../../utilities/semantic-types/tuple';
 import { buildEstinant } from '../../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../../adapter/odeshin/odeshin2';
 import {
   TYPE_SCRIPT_FILE_IMPORT_LIST_GEPP,
   TypeScriptFileImportListVoque,
