+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/file/getBoundedFile.ts
@@ -1,7 +1,7 @@
 import { assertNotNull } from '../../../../utilities/assertNotNull';
 import { isNotNull } from '../../../../utilities/isNotNull';
 import { buildEstinant } from '../../../../adapter/estinant-builder/estinantBuilder';
-import { OdeshinZorn } from '../../../../adapter/odeshin2';
+import { OdeshinZorn } from '../../../../adapter/odeshin/odeshin2';
 import {
   FILE_ANCESTOR_DIRECTORY_PATH_SET_GEPP,
   FileAncestorDirectoryPathSetVoque,
