+++ packages/voictents-and-estinants-engine/src/adapted-programs/programs/render-knowledge-graph/renderKnowledgeGraph.ts
@@ -3,7 +3,7 @@ import {
   buildGeppCombination,
   buildVoictentByGepp,
   digikikify,
-} from '../../../adapter/digikikify';
+} from '../../../adapter/engine/digikikify';
 import { ProgramFileCache } from '../../../utilities/programFileCache';
 import {
   EngineFunctionConfigurationVoque,
