+++ packages/voictents-and-estinants-engine/src/adapter/engine/digikikify.ts
@@ -2,51 +2,51 @@ import { Simplify, UnionToIntersection } from 'type-fest';
 import {
   DigikikifierInput as CoreDigikikifierInput,
   digikikify as coreDigikikify,
-} from '../core/engine/digikikify';
+} from '../../core/engine/digikikify';
 import {
   GenericEstinant2Tuple,
   UnsafeEstinant2,
   GenericEstinant2,
   UnsafeEstinant2Tuple,
   Estinant2,
-} from '../core/engine-shell/estinant/estinant';
+} from '../../core/engine-shell/estinant/estinant';
 import {
   GenericLeftInputVicken,
   LeftInputVicken,
-} from '../core/engine-shell/vicken/leftInputVicken';
+} from '../../core/engine-shell/vicken/leftInputVicken';
 import {
   GenericOutputVicken,
   OutputVicken,
-} from '../core/engine-shell/vicken/outputVicken';
+} from '../../core/engine-shell/vicken/outputVicken';
 import {
   GenericRightInputVickenTuple,
   RightInputVicken,
-} from '../core/engine-shell/vicken/rightInputVicken';
+} from '../../core/engine-shell/vicken/rightInputVicken';
 import {
   Gepp,
   GenericGeppCombination,
   GeppTuple,
-} from '../core/engine-shell/voictent/gepp';
+} from '../../core/engine-shell/voictent/gepp';
 import {
   GenericInMemoryOdeshin2ListVoque,
   InMemoryOdeshin2ListVoictent,
-} from '../core/engine/inMemoryOdeshinVoictent2';
+} from '../../core/engine/inMemoryOdeshinVoictent2';
 import {
   GenericVoictent2,
   GenericVoictent2Tuple,
   UnsafeVoictent2Tuple,
   Voictent2,
-} from '../core/engine/voictent2';
-import { GenericVoque } from '../core/engine/voque';
-import { ProgramErrorGepp } from '../adapted-programs/programmable-units/error/programError';
-import { GenericAbstractSerializableSourceVoque } from '../core-programs/abstractSerializableVoictent';
-import { buildAddMetadataForSerialization } from '../core-programs/buildAddMetadataForSerialization';
-import { SerializableVoictent } from '../core-programs/serializableVoictent';
-import { ProgramFileCache } from '../utilities/programFileCache';
+} from '../../core/engine/voictent2';
+import { GenericVoque } from '../../core/engine/voque';
+import { ProgramErrorGepp } from '../../adapted-programs/programmable-units/error/programError';
+import { GenericAbstractSerializableSourceVoque } from '../../core-programs/abstractSerializableVoictent';
+import { buildAddMetadataForSerialization } from '../../core-programs/buildAddMetadataForSerialization';
+import { SerializableVoictent } from '../../core-programs/serializableVoictent';
+import { ProgramFileCache } from '../../utilities/programFileCache';
 import {
   FileSystemNodeVoictent,
   GenericFileSystemNodeVoque,
-} from '../adapted-programs/programmable-units/file/fileSystemNodeVoictent';
+} from '../../adapted-programs/programmable-units/file/fileSystemNodeVoictent';
 
 type VoqueUnionFromVoictentTuple<
   TVoictentTuple extends UnsafeVoictent2Tuple,
